refactor AppSettings to use useAppTranslation hook and group settings in accordion
